# Conflicts:
#	package/crm/UPGRADE-1.10.md
#	package/crm/src/OroCRM/Bundle/SalesBundle/Resources/config/services.yml

+++ src/Oro/Bundle/ImapBundle/Sync/ImapEmailSynchronizationProcessor.php
@@ -41,6 +41,9 @@ class ImapEmailSynchronizationProcessor extends AbstractEmailSynchronizationProc
     /** Determines how often the clearing of outdated folders routine should be executed */
     const CLEANUP_EVERY_N_RUN = 100;
 
+    /** Time limit to sync origin in seconds */
+    const MAX_ORIGIN_SYNC_TIME = 30;
+
     /** @var ImapEmailManager */
     protected $manager;
 
@@ -71,7 +74,7 @@ class ImapEmailSynchronizationProcessor extends AbstractEmailSynchronizationProc
         $this->emailEntityBuilder->clear();
 
         $this->initEnv($origin);
-
+        $processStartTime = time();
         // iterate through all folders enabled for sync and do a synchronization of emails for each one
         $imapFolders = $this->getSyncEnabledImapFolders($origin);
         foreach ($imapFolders as $imapFolder) {
@@ -103,6 +106,11 @@ class ImapEmailSynchronizationProcessor extends AbstractEmailSynchronizationProc
             }
 
             $this->cleanUp(true, $imapFolder->getFolder());
+
+            $processSpentTime = time() - $processStartTime;
+            if ($processSpentTime > self::MAX_ORIGIN_SYNC_TIME) {
+                break;
+            }
         }
 
         $this->removeRemotelyRemovedEmails($origin);
@@ -148,12 +156,18 @@ class ImapEmailSynchronizationProcessor extends AbstractEmailSynchronizationProc
                     /* @var $staleImapEmails ImapEmail[] */
                     foreach ($staleImapEmails as $imapEmail) {
                         $email = $imapEmail->getEmail();
-                        $email->getEmailUsers()->forAll(function ($key, EmailUser $emailUser) use ($folder) {
-                            $emailUser->removeFolder($folder);
-                            if (!$emailUser->getFolders()->count()) {
-                                $this->em->remove($emailUser);
-                            }
-                        });
+                        $email->getEmailUsers()
+                            ->forAll(function ($key, EmailUser $emailUser) use ($folder, $imapEmail) {
+                                $existsEmails = $this->em->getRepository('OroImapBundle:ImapEmail')
+                                    ->findBy(['email' => $imapEmail->getEmail()]);
+
+                                $emailUser->removeFolder($folder);
+                                // if existing imapEmail is last for current email or is absent
+                                // we remove emailUser and after that will remove last imapEmail and email
+                                if (count($existsEmails) <= 1 && !$emailUser->getFolders()->count()) {
+                                    $this->em->remove($emailUser);
+                                }
+                            });
                         $this->em->remove($imapEmail);
                     }
                 });
@@ -563,7 +577,10 @@ class ImapEmailSynchronizationProcessor extends AbstractEmailSynchronizationProc
 
         $emailUser = $imapEmail->getEmail()->getEmailUserByFolder($imapEmail->getImapFolder()->getFolder());
         if ($emailUser != null) {
-            $imapEmail->getEmail()->getEmailUsers()->removeElement($emailUser);
+            $emailUser->removeFolder($imapEmail->getImapFolder()->getFolder());
+            if (!$emailUser->getFolders()->count()) {
+                $imapEmail->getEmail()->getEmailUsers()->removeElement($emailUser);
+            }
         }
         $this->em->remove($imapEmail);
     }
